Program Management updates #2
Added a check for .page-title elements before getProgramTitle() runs.
Created a function to dynamically update the breadcrumb link text so it
matches the Title of the current page.
Reworked Title conditionals.
Added an extra check in recordLogin.js to check for the target element
before it runs showGettingStarted()
Updated the URL in program-list.mustache.html to face the new Program
Mgmt Marcom page.