Get User Configurations Updates
Formatted program-list.mustache.html and moved the <a> with the title
up one level in the dom.
Made progress on getUserConfigurations.js, error handling,
documentation, better url # checking (still not very robust), changed
ajax request type to GET, removed
customCheckAndRadioBoxes.customCheckbox() because a different script
checks for it and loads it.

Working on a solution for passing the programId around, started with
using getProgramParticipationStats.js as a starting point.